refactor and export initAll() method
