+++ src/demo/index.jsx
@@ -1,19 +1,23 @@
 import 'babel-polyfill';
 import $ from 'jquery';
-import ReactDOM from 'react-dom';
-import React from 'react';
-import UI from './ui';
-import constants, {Modes} from './constants';
-import {setInitialState, setSetStateCallback} from './state';
-import {render as renderCanvas} from './renderer';
 import {queryStrFor} from './helpers';
-import {toMode} from './toMode';
+import {initAll} from './init';
 
 let currentAppMode = queryStrFor('mode') || 'intro-instructions';
+let canvas, backgroundCanvas;
 
 function onLevelChange(event) {
   currentAppMode = event.target.id;
-  initAll(currentAppMode);
+  initDemoPage();
+}
+
+function initDemoPage() {
+  initAll({
+    appMode: currentAppMode,
+    onContinue,
+    canvas,
+    backgroundCanvas
+  });
 }
 
 function onContinue() {
@@ -29,44 +33,12 @@ function onContinue() {
 }
 
 $(document).ready(() => {
+  // Set up canvases.
+  canvas = document.getElementById('activity-canvas');
+  backgroundCanvas = document.getElementById('background-canvas');
+
   $(`#${currentAppMode}`).prop('checked', true);
   $('.level-radio').change(onLevelChange);
 
-  initAll(currentAppMode);
+  initDemoPage();
 });
-
-export const initAll = function(appMode) {
-  // Set up canvases.
-  const canvas = document.getElementById('activity-canvas');
-  const backgroundCanvas = document.getElementById('background-canvas');
-  canvas.width = backgroundCanvas.width = constants.canvasWidth;
-  canvas.height = backgroundCanvas.height = constants.canvasHeight;
-
-  // Set initial state for UI elements.
-  setInitialState({
-    appMode,
-    canvas,
-    backgroundCanvas,
-    currentMode: Modes.Loading,
-    onContinue
-  });
-
-  // Initialize our first model.
-  toMode(Modes.Loading);
-
-  // Start the canvas renderer.  It will self-perpetute by calling
-  // requestAnimationFrame on itself.
-  renderCanvas();
-
-  // Render the UI.
-  renderUI();
-
-  // And have the render UI handler be called every time state is set.
-  setSetStateCallback(renderUI);
-};
-
-// Tell React to explicitly render the UI.
-export const renderUI = () => {
-  const renderElement = document.getElementById('container-react');
-  ReactDOM.render(<UI />, renderElement);
-};
